rename interfaces folder
