+++ apps/server/src/auth/auth.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { User, UsersService } from '../users/users.service';
 import { JwtService } from '@nestjs/jwt';
-import { JwtPayload } from './Interfaces/JwtPayload';
-import { AuthEntity } from './Entities/auth.entity';
+import { JwtPayload } from './interfaces/jwt-payload';
+import { AuthEntity } from './entities/auth.entity';
 
 @Injectable()
 export class AuthService {
